requirements.txt
altair==4.2.0
numpy==1.21.5
pandas==1.4.2
Pillow==9.2.0
plotly==5.6.0
seaborn==0.11.2
streamlit==1.12.2
